templates/ravenmaskos/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ravenmaskos"
version = "0.1.0"
description = "Backend service template with FastAPI, PostgreSQL, Redis, Kafka, and observability"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Your Name", email = "you@example.com" }
]

dependencies = [
    # API & Models
    "fastapi>=0.109.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Runtime
    "uvicorn[standard]>=0.27.0",

    # PostgreSQL
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.25",

    # Redis
    "redis>=5.0.0",

    # Kafka
    "aiokafka>=0.10.0",

    # Security / Crypto
    "authlib>=1.3.0",
    "python-jose[cryptography]>=3.3.0",
    "argon2-cffi>=23.1.0",
    "bcrypt>=4.1.0",

    # AWS Integration
    "boto3>=1.34.0",
    "aioboto3>=12.0.0",

    # Observability
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-instrumentation-asyncpg>=0.43b0",
    "opentelemetry-instrumentation-redis>=0.43b0",
    "prometheus-client>=0.19.0",

    # Logging
    "structlog>=24.1.0",

    # Rate Limiting
    "slowapi>=0.1.9",

    # CloudEvents
    "cloudevents>=1.10.0",

    # Database Migrations
    "alembic>=1.13.0",

    # Authorization
    "openfga-sdk>=0.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
start = "uvicorn src.app.main:app --reload"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src/app --cov-report=term-missing"
